backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.27
pydantic>=2.6.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
google-generativeai>=0.4.0
python-multipart>=0.0.9
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pytest>=8.0.0
